feat: Implement classification model client-side
This commit scaffolds the basic application we will use across our
experiments. The UI consists of an input field through which users
can enter the text that will be analyzed by the TF Toxicity model.
The result of the model's classification, will be displayed back
to the user via the same UI.

In this commit, both the model loading and the classification happen
on the client side. The model is loaded once the page is loaded, for
better performance. While performing the input text sentiment
analysis, we internally keep track of two metrics:
- the elapsed time from when the input text was submitted for
analysis, to when the result is returned by the classifier
- the average of the above value across the number of user text
submissions

The above metrics will remain consistent across all our experiments,
in order to correctly evaluate which implementation has better
performance.