feat: Implement classification model in Worker
This commit moves the TF model implementation from the client
to a Cloudflare Worker (a Pages Function to be more precise).
This includes both the model loading and the classification
logic.

With these changes in place, once a user submits their text
for sentiment analysis, a request is made to the `/classify`
endpoint. The aforementioned route is matched by our Pages
Function. This Function, will:
- load the TF model (when called for the first time)
- run the prediction on the given user input
- return the prediction result to the client

It is important to note that the metrics the client keeps
track of, more specifically the classification elapsed time,
will now include the duration of the round-trip XHR request
to the Worker, **and** the duration of the classification
itself. The classification elapsed time metric alone is
kept track of within the Worker, and can be accessed by
monitoring the Worker's logs (or the terminal logs, if
application is run locally).